+++ CHANGELOG.md
@@ -29,7 +29,7 @@
 ### New Filter and Binding Features
 
 - Added `ISignalFilter` interface for custom signal filtering.
-- Updated `Publish` method to accept filters.
+- Updated `Emit` method to accept filters.
 - New Binding options.
 
 ## [1.0.3] - 2025-09-02